// hr-frontend/src/lib/api/generated/model/employmentOverview.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface EmploymentOverview {
  employeeId?: number;
  employeeNumber?: string;
  hireDate?: string;
  startDate?: string;
  terminationDate?: string;
  employeeStatusId?: number;
  statusCode?: string;
  statusName?: string;
  statusNameAr?: string;
  organizationId?: number;
  organizationCode?: string;
  organizationNameEn?: string;
  organizationNameAr?: string;
  orgUnitId?: number;
  orgUnitCode?: string;
  orgUnitName?: string;
  orgUnitNameAr?: string;
  positionAssignmentId?: number;
  assignmentType?: number;
  assignmentTypeName?: string;
  assignmentStartDate?: string;
  assignmentEndDate?: string;
  positionId?: number;
  positionCode?: string;
  positionTitle?: string;
  positionTitleAr?: string;
  gradeId?: number;
  gradeCode?: string;
  gradeName?: string;
  gradeNameAr?: string;
  gradeRank?: number;
  teamLeaderEmployeeId?: number;
  teamLeaderName?: string;
  managerEmployeeId?: number;
  managerName?: string;
  reportsToPositionId?: number;
  reportsToPositionTitle?: string;
  reportsToEmployeeId?: number;
  reportsToEmployeeName?: string;
}